funcionalidad de la base de datos asi como primeras llamadas a la base de datos desde swagger-ui

+++ fruteriaSpring/src/main/java/fruteria/Repository/MarcaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fruteria.Entity.MarcaEntity;
 
-public interface MarcaRepository  extends JpaRepository<MarcaEntity, Integer>{
+public interface MarcaRepository  extends JpaRepository<MarcaEntity, Long>{
 
 }
